(minor correction) global_state -> self.global_state
