+++ src/appfl/algorithm/Client_sgd.py
@@ -1,0 +1,67 @@
+import logging
+
+log = logging.getLogger(__name__)
+
+from collections import OrderedDict
+from .algorithm import BaseServer, BaseClient
+
+import torch
+from torch.optim import *
+from torch.nn import CrossEntropyLoss
+from torch.utils.data import DataLoader
+import copy
+
+ 
+class ClientSGD(BaseClient):
+    def __init__(self, id, weight, model, dataloader, device, **kwargs):
+        super(ClientSGD, self).__init__(id, weight, model, dataloader, device)
+        self.__dict__.update(kwargs)
+        self.loss_fn = CrossEntropyLoss()
+
+    def update(self):
+
+        self.model.train()
+        self.model.to(self.device)
+
+        optimizer = eval(self.optim)(self.model.parameters(), **self.optim_args)
+
+        """ Inputs for the local model update """
+        ## "global_state" from a server is already stored in 'self.model'
+
+        """ Multiple local update """
+        for i in range(self.num_local_epochs):
+            for data, target in self.dataloader:
+
+                data = data.to(self.device)
+                target = target.to(self.device)
+
+                optimizer.zero_grad()
+                output = self.model(data)
+                loss = self.loss_fn(output, target)
+                loss.backward()
+
+                if self.clip_value != False:                                              
+                    torch.nn.utils.clip_grad_norm_(self.model.parameters(), self.clip_value, norm_type=self.clip_norm)                
+
+                optimizer.step()
+
+        self.primal_state = copy.deepcopy(self.model.state_dict()) 
+
+        """ Differential Privacy  """
+        if self.epsilon != False:
+            sensitivity = 0
+            if self.clip_value != False:                           
+                sensitivity = 2.0 * self.clip_value * self.optim_args.lr 
+            scale_value = sensitivity / self.epsilon            
+            super(ClientSGD, self).laplace_mechanism_output_perturb(scale_value)
+
+        """ Update local_state """
+        self.local_state = OrderedDict()
+        self.local_state["primal"] = copy.deepcopy(self.primal_state)
+        self.local_state["dual"] = OrderedDict()
+        self.local_state["penalty"] = OrderedDict()
+        self.local_state["penalty"][self.id] = 0.0
+
+        return self.local_state
+
+
