+++ src/appfl/algorithm/Server_fed_avg.py
@@ -1,6 +1,6 @@
-from .fedadapt import FedAdaptServer
+from .Server_fed import FedServer
 
-class FedAvgServer(FedAdaptServer):
+class ServerFedAvg(FedServer):
     def compute_step(self):        
         for name, _ in self.model.named_parameters():                        
             self.step[name] = self.pseudo_grad[name]
