+++ src/appfl/algorithm/Server_fed_yogi.py
@@ -1,4 +1,4 @@
-from .fedadapt import FedAdaptServer
+from .Server_fed import FedAdaptServer
 import torch
 
 class FedYogiServer(FedAdaptServer):
