+++ src/appfl/config/fed/federated.py
@@ -1,16 +1,37 @@
 from dataclasses import dataclass
 from omegaconf import DictConfig, OmegaConf
 
+""" various server updates are available 
+(Baseline)
+    ServerFedAvg
+(Group 1)
+    ServerFedAvgMomentum
+    ServerFedAdagrad
+    ServerFedAdam 
+    ServerFedYogi
+(Group 2)
+    ServerFedBFGS
+    ServerFedBroyden
+    ServerFedDFP
+    ServerFedSR1
+""" 
 @dataclass
-class FedAvg:
-    type: str = "fedavg"
-    servername: str = "FedAvgServer"
-    clientname: str = "FedAvgClient"
+class Federated:
+    type: str = "federated"  
+    servername: str = "ServerFedAvg" 
+    clientname: str = "ClientSGD"
     args: DictConfig = OmegaConf.create(
-        {
-            "num_local_epochs": 10,
+        {           
 
-            ## Optimizer
+            ## Server update
+            "server_learning_rate": 0.01,
+
+            "server_adapt_param": 0.001,
+            "server_momentum_param_1": 0.9,
+            "server_momentum_param_2": 0.99,
+
+            ## Clients update
+            "num_local_epochs": 10,
             "optim": "SGD",
             "optim_args": {
                 "lr": 0.01,
